+++ CHANGELOG.md
@@ -14,6 +14,12 @@ changes.
 
 - HeadLogic Outcome is now being trace on every protocol step transition.
 
+- **BREAKING** `hydra-cardano-api` changes:
+  + Remove `Hydra.Cardano.Api.SlotNo` module.
+  + Replace `fromConsensusPointHF` with `fromConsensusPointInMode` and
+    `toConsensusPointHF` with `toConsensusPointInMode`.
+  + Re-export new `AcquiringFailure` type from `cardano-api`.
+
 ## [0.8.1] - 2022-11-17
 
 - **BREAKING** Implemented [ADR18](https://hydra.family/head-protocol/adr/18) to keep only a single state: